feat/refactor data and clean up docs
