+++ api/package.json
@@ -9,8 +9,8 @@
 		"dev": "bun --env-file=../.env --watch src/index.ts",
 		"build": "bun build src/index.ts --outdir dist --target bun",
 		"start": "bun run db:migrate && bun --env-file=../.env src/index.ts",
-		"health": "bun src/readiness.ts",
-		"health:seeded": "bun src/readiness.ts --require-seeded",
+		"health": "bun src/index.ts --health",
+		"health:seeded": "bun src/index.ts --health --require-seeded",
 		"test:unit": "bun --env-file=../.env test src/",
 		"test:e2e": "bun --env-file=../.env test e2e/",
 		"lint": "bunx biome check",
