--- api/src/broadcast.ts
@@ -1,1 +1,0 @@
-export type BroadcastFn = (eventName: string, payload: unknown) => void;
